[Pacage CMO] Update and add more tests
